refactor: adjust project directories
