+++ backend/src/models/kanban/JiraBoard/JiraCard.ts
