+++ backend/src/services/codebase/GitHub/GitHub.ts
@@ -1,13 +1,13 @@
-import { Codebase } from "./Codebase";
-import { DeployTimes } from "../../models/pipeline/DeployTimes";
+import { Codebase } from "../Codebase";
+import { DeployTimes } from "../../../models/pipeline/DeployTimes";
 import axios, { AxiosInstance } from "axios";
 import { JsonConvert } from "json2typescript";
 import GitUrlParse from "git-url-parse";
-import { GitHubPull } from "../../models/codebase/GitHubPull";
-import { CommitInfo } from "../../models/codebase/CommitInfo";
-import { LeadTime, PipelineLeadTime } from "../../models/codebase/LeadTime";
-import { GitOrganization } from "../../models/codebase/GitOrganization";
-import { GitHubRepo } from "../../models/codebase/GitHubRepo";
+import { GitHubPull } from "../../../models/codebase/GitHub/GitHubPull";
+import { CommitInfo } from "../../../models/codebase/CommitInfo";
+import { LeadTime, PipelineLeadTime } from "../../../models/codebase/LeadTime";
+import { GitOrganization } from "../../../models/codebase/GitOrganization";
+import { GitHubRepo } from "../../../models/codebase/GitHub/GitHubRepo";
 
 export class GitHub implements Codebase {
   private httpClient: AxiosInstance;
@@ -21,7 +21,7 @@ export class GitHub implements Codebase {
       },
     });
   }
-  
+
   async fetchAllOrganization(): Promise<string[]> {
     const response = await this.httpClient.get("/user/orgs");
     const gitOrganizations = new JsonConvert().deserializeArray(
@@ -29,30 +29,32 @@ export class GitHub implements Codebase {
       GitOrganization
     );
     return gitOrganizations.map((org) => org.orgName);
-   }
+  }
 
   async fetchAllRepo(gitOrganizations: string[]): Promise<string[]> {
     const result: string[] = [];
     const requestUrl: string[] = [];
-    
+
     requestUrl.push("user/repos");
-    gitOrganizations.forEach(organization => { requestUrl.push(`/orgs/${organization}/repos`); });
-    
-    await Promise.all(requestUrl.map(
-      async (url) => {
+    gitOrganizations.forEach((organization) => {
+      requestUrl.push(`/orgs/${organization}/repos`);
+    });
+
+    await Promise.all(
+      requestUrl.map(async (url) => {
         const response = await this.httpClient.get(url);
         const gitHubRepos: GitHubRepo[] = new JsonConvert().deserializeArray(
           response.data,
           GitHubRepo
-          );
-          gitHubRepos.forEach((repo) =>{
-            if(!result.includes(repo.repoUrl)) {
-              result.push(repo.repoUrl);
-            }
-          });
-      }
-    ));
-    
+        );
+        gitHubRepos.forEach((repo) => {
+          if (!result.includes(repo.repoUrl)) {
+            result.push(repo.repoUrl);
+          }
+        });
+      })
+    );
+
     return result;
   }
 
@@ -81,14 +83,16 @@ export class GitHub implements Codebase {
               const response = await this.httpClient.get(
                 `/repos/${repository}/commits/${deployInfo.commitId}/pulls`
               );
-              
-              const gitHubPulls: GitHubPull[] = new JsonConvert().deserializeArray(
-                response.data,
-                GitHubPull
-              );
-              
-              const jobFinishTime: number = new Date(deployInfo.jobFinishTime).getTime();
-              const pipelineCreateTime: number = new Date(deployInfo.pipelineCreateTime).getTime();
+
+              const gitHubPulls: GitHubPull[] =
+                new JsonConvert().deserializeArray(response.data, GitHubPull);
+
+              const jobFinishTime: number = new Date(
+                deployInfo.jobFinishTime
+              ).getTime();
+              const pipelineCreateTime: number = new Date(
+                deployInfo.pipelineCreateTime
+              ).getTime();
 
               const noMergeDelayTime = new LeadTime(
                 deployInfo.commitId,
@@ -101,11 +105,11 @@ export class GitHub implements Codebase {
               if (gitHubPulls.length == 0) {
                 return noMergeDelayTime;
               }
-              
+
               const mergedPull: GitHubPull | undefined = gitHubPulls
                 .filter((gitHubPull) => gitHubPull.mergedAt != null)
                 .pop();
-              
+
               if (mergedPull == undefined) {
                 return noMergeDelayTime;
               }
@@ -114,11 +118,9 @@ export class GitHub implements Codebase {
               const prResponse = await this.httpClient.get(
                 `/repos/${repository}/pulls/${mergedPull.number}/commits`
               );
-              
-              const gitHubCommites: CommitInfo[] = new JsonConvert().deserializeArray(
-                prResponse.data,
-                CommitInfo
-              );
+
+              const gitHubCommites: CommitInfo[] =
+                new JsonConvert().deserializeArray(prResponse.data, CommitInfo);
 
               //get the first commit.
               const firstCommit: CommitInfo = gitHubCommites[0];
