--- backend/src/types/ThereIsNoCardsInDoneColumn.ts
@@ -1,1 +1,0 @@
-export class ThereIsNoCardsInDoneColumn extends Error {}
